Update pfod.cpp
"Charge start" slider needs to be able to go a bit over the Full voltage of the battery.  Otherwise if you touch the slider, you can now only set it to the max = battFull. If you try to charge manually and the charger gives out the same voltage as batFul, it will newer start charging.  So need to be able to set it a bit over batFull.